etc: organize packages
