--- authorization-server/src/main/java/com/woowahan/techcourse/authorizationserver/exception/GithubApiFailException.java
@@ -1,5 +1,0 @@
-package com.woowahan.techcourse.authorizationserver.exception;
-
-public class GithubApiFailException extends RuntimeException {
-
-}
