+++ authorization-server/src/main/java/com/woowahan/techcourse/authorizationserver/exception/github/GithubApiFailException.java
@@ -1,0 +1,5 @@
+package com.woowahan.techcourse.authorizationserver.exception.github;
+
+public class GithubApiFailException extends RuntimeException {
+
+}
