+++ authorization-server/src/main/java/com/woowahan/techcourse/authorizationserver/exception/github/GithubConnectionException.java
@@ -1,4 +1,4 @@
-package com.woowahan.techcourse.authorizationserver.exception;
+package com.woowahan.techcourse.authorizationserver.exception.github;
 
 public class GithubConnectionException extends RuntimeException {
